Moves trampoline pattern to java 11

+++ trampoline/src/test/java/com/iluwatar/trampoline/TrampolineAppTest.java
@@ -23,22 +23,20 @@
 
 package com.iluwatar.trampoline;
 
+import static org.junit.Assert.assertEquals;
+
 import org.junit.Test;
-
-import java.io.IOException;
-
-import static org.junit.Assert.*;
 
 
 /**
  * Test for trampoline pattern.
- * */
+ */
 public class TrampolineAppTest {
 
 
   @Test
   public void testTrampolineWithFactorialFunction() {
-    int result = TrampolineApp.loop(10, 1).result();
+    long result = TrampolineApp.loop(10, 1).result();
     assertEquals("Be equal", 3628800, result);
   }
